refactor: group inventory and modular components
