+++ src/App.tsx
@@ -80,15 +80,15 @@ import { notify } from "./lib/notify";
 import { collectNewMatches } from "./fissureAlerts";
 import Statistics from "./Statistics";
 import Overlay from "./Overlay";
-import ModularWindow from "./ModularWindow";
-import ModularWindowPage from "./ModularWindowPage";
+import ModularWindow from "./modular-window/ModularWindow";
+import ModularWindowPage from "./modular-window/ModularWindowPage";
 import SettingsModal from "./SettingsModal";
 import ChangeLog from "./ChangeLog";
-import InventoryGrid from "./InventoryGrid";
-import InventoryBatchPreview from "./InventoryBatchPreview";
-import InventoryToolbar from "./InventoryToolbar";
+import InventoryGrid from "./inventory/InventoryGrid";
+import InventoryBatchPreview from "./inventory/InventoryBatchPreview";
+import InventoryToolbar from "./inventory/InventoryToolbar";
 import AppNavigation, { type Module } from "./AppNavigation";
-import InventorySidebar from "./InventorySidebar";
+import InventorySidebar from "./inventory/InventorySidebar";
 import CompletionistTabs, { type CompletionistView } from "./CompletionistTabs";
 import HeaderActions from "./HeaderActions";
 import ErrorBoundary from "./shared/ErrorBoundary";
@@ -124,7 +124,6 @@ import type { TradeCompletedEvent } from "./types/trades";
 import type { AddTradeArgs, AnalyzeRivenArgs, BlobStatusPayload, InventoryRewardPayload, ItemListStatus, OcrRivenScreenResult, OverlayWindowBounds, RelicRewardsPayload, SaveApiInventoryArgs, SavedApiInventory, SettingsFile, SettingsPatch, WarframeCredentials, WarframeInventoryRequest, WarframeWindowRect, WfmCredentials, WfmSession } from "./types/tauri";
 import "./App.css";
 import "./images.css";
-import "./InventoryGrid.css";
 
 const _winLabel = getCurrentWindow().label;
 // Support all URL formats: query string (?overlay), hash (#overlay), or window label.
