+++ src/inventory/InventoryBatchPreview.css
