+++ src/inventory/InventoryBatchPreview.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useRef, useState } from "react";
 import InventoryGrid from "./InventoryGrid";
-import { ViewToggle } from "./shared/ViewToggle";
-import type { ViewMode } from "./types/ui";
+import { ViewToggle } from "../shared/ViewToggle";
+import type { ViewMode } from "../types/ui";
 import "./InventoryBatchPreview.css";
 
 interface InventoryBatchPreviewProps {
