+++ src/inventory/InventoryGrid.css
