+++ src/inventory/InventorySidebar.tsx
